Grades: move out of temporary new folder EDUCATOR-1400
