+++ lms/djangoapps/certificates/management/commands/fix_ungraded_certs.py
@@ -8,7 +8,7 @@ from django.core.management.base import BaseCommand
 
 from certificates.models import GeneratedCertificate
 from courseware import courses
-from lms.djangoapps.grades.new.course_grade_factory import CourseGradeFactory
+from lms.djangoapps.grades.course_grade_factory import CourseGradeFactory
 
 log = logging.getLogger(__name__)
