+++ lms/djangoapps/certificates/tests/test_signals.py
@@ -11,7 +11,7 @@ from certificates.models import (
     GeneratedCertificate,
     CertificateStatuses,
 )
-from lms.djangoapps.grades.new.course_grade_factory import CourseGradeFactory
+from lms.djangoapps.grades.course_grade_factory import CourseGradeFactory
 from lms.djangoapps.grades.tests.utils import mock_passing_grade
 from lms.djangoapps.verify_student.models import SoftwareSecurePhotoVerification
 from openedx.core.djangoapps.certificates.config import waffle
