+++ lms/djangoapps/grades/tests/utils.py
@@ -16,8 +16,8 @@ def mock_passing_grade(grade_pass='Pass', percent=0.75, ):
     """
     Mock the grading function to always return a passing grade.
     """
-    with patch('lms.djangoapps.grades.new.course_grade.CourseGrade._compute_letter_grade') as mock_letter_grade:
-        with patch('lms.djangoapps.grades.new.course_grade.CourseGrade._compute_percent') as mock_percent_grade:
+    with patch('lms.djangoapps.grades.course_grade.CourseGrade._compute_letter_grade') as mock_letter_grade:
+        with patch('lms.djangoapps.grades.course_grade.CourseGrade._compute_percent') as mock_percent_grade:
             mock_letter_grade.return_value = grade_pass
             mock_percent_grade.return_value = percent
             yield
@@ -28,7 +28,7 @@ def mock_get_score(earned=0, possible=1, first_attempted=datetime(2000, 1, 1, 0,
     """
     Mocks the get_score function to return a valid grade.
     """
-    with patch('lms.djangoapps.grades.new.subsection_grade.get_score') as mock_score:
+    with patch('lms.djangoapps.grades.subsection_grade.get_score') as mock_score:
         mock_score.return_value = ProblemScore(
             raw_earned=earned,
             raw_possible=possible,
