+++ lms/djangoapps/instructor/views/gradebook_api.py
@@ -12,7 +12,7 @@ from opaque_keys.edx.keys import CourseKey
 
 from courseware.courses import get_course_with_access
 from edxmako.shortcuts import render_to_response
-from lms.djangoapps.grades.new.course_grade_factory import CourseGradeFactory
+from lms.djangoapps.grades.course_grade_factory import CourseGradeFactory
 from lms.djangoapps.instructor.views.api import require_level
 from xmodule.modulestore.django import modulestore
